Create validate request test, move openai request test, add body to lambda test
