+++ test/unit/test_utils.py
@@ -1,0 +1,33 @@
+import pytest
+from unittest.mock import patch
+
+from src.lambda_src.utils import get_openai_model_response, validate_request_body
+
+FILEPATH = "src.lambda_src.utils"
+
+
+@pytest.mark.parametrize(
+    ("request_body", "expected"),
+    [
+        ([], False),
+        ([{"image_path": "blah"}], False),
+        ([{"image_path": "blah", "image_caption": "blah"}], True)
+    ]
+)
+def test_validate_request_body(request_body, expected):
+    assert validate_request_body(request_body) == expected
+
+
+@patch(f"{FILEPATH}.requests.Response")
+@patch(f"{FILEPATH}.requests")
+def test_get_openai_response(mock_requests, mock_response, sample_openai_response):
+    mock_response.json.return_value = sample_openai_response
+    mock_requests.post.return_value = mock_response
+
+    response = get_openai_model_response(image_captions=[])
+
+    assert type(response) == list
+
+    expected_keys = ["image_path", "image_caption", "story"]
+    response_body_has_expected_keys = all(set(expected_keys) == set(item.keys()) for item in response)
+    assert response_body_has_expected_keys
